Refactor front-end directory to primarily group by resource
This keeps related files closer together, making the directory a lot easier to navigate in general without having to keep bouncing between different folders.
