+++ apps/front-end/src/components/Navigation/index.tsx
@@ -2,8 +2,8 @@ import type { ReactNode } from "react";
 
 import { useScreenSize } from "@alextheman/components";
 
-import NavigationBottom from "src/Navigation/NavigationBottom";
-import NavigationDrawer from "src/Navigation/NavigationDrawer";
+import NavigationBottom from "src/components/Navigation/NavigationBottom";
+import NavigationDrawer from "src/components/Navigation/NavigationDrawer";
 
 export interface NavigationProps {
   children: ReactNode;
