+++ apps/front-end/src/resources/Albums/pages/mutations/CreateAlbum.tsx
@@ -3,9 +3,9 @@ import type { AlbumToPost } from "@neurosongs/types";
 import { useSnackbar } from "@alextheman/components";
 import { useNavigate } from "react-router-dom";
 
-import AlbumForm from "src/components/resources/albums/AlbumForm";
 import formatError from "src/errors/formatError";
-import { useCreateAlbumMutation } from "src/queries/albums";
+import { useCreateAlbumMutation } from "src/resources/Albums/album-queries";
+import AlbumForm from "src/resources/Albums/components/AlbumForm";
 
 function CreateAlbum() {
   const userId = "a05cadd9-479c-43f7-a86f-487e38a282f7";
