+++ apps/front-end/src/resources/Users/pages/User/index.tsx
@@ -12,9 +12,9 @@ import { useParams } from "react-router-dom";
 
 import LoaderProvider from "src/components/LoaderProvider";
 import neurosongsNote from "src/images/Neurosongs_note.png";
-import UserAlbums from "src/pages/Users/User/UserAlbums";
-import UserSongs from "src/pages/Users/User/UserSongs";
-import { useUserQuery } from "src/queries/users";
+import UserAlbums from "src/resources/Users/pages/User/UserAlbums";
+import UserSongs from "src/resources/Users/pages/User/UserSongs";
+import { useUserQuery } from "src/resources/Users/user-queries";
 
 type TabState = "songs" | "albums";
