+++ apps/front-end/src/resources/Users/user-queries.ts
@@ -3,7 +3,7 @@ import type { PublicUser } from "@neurosongs/types";
 import { parsePublicUser } from "@neurosongs/types";
 import { useQuery } from "@tanstack/react-query";
 
-import neurosongsAxiosClient from "src/queries/neurosongsAxiosClient";
+import neurosongsAxiosClient from "src/neurosongsAxiosClient";
 
 export function useUserQuery(userId: string) {
   return useQuery<PublicUser>({
